requirements.txt
streamlit
python-dotenv
beautifulsoup4
requests
pypdf
sentence-transformers
faiss-cpu
groq
fastapi "uvicorn[standard]"